+++ frontend/.env
@@ -1,3 +1,4 @@
 SKIP_PREFLIGHT_CHECK=true
 
 REACT_APP_API_URL=http://15.164.68.136:8080/
+//process.env.
