+++ frontend/src/components/atomic/ItemPrice.jsx
@@ -1,29 +1,34 @@
 import styled from 'styled-components/';
 import { theme } from '../style/Theme';
 
+function ItemPrice({ sPrice, nPrice, type }) {
+	if (sPrice) sPrice = priceComma(sPrice);
+	nPrice = priceComma(nPrice);
+	return type === 'basic' ? (
+		<>
+			<ItemPriceBlock>{sPrice ? sPrice : nPrice}원</ItemPriceBlock>
+			{sPrice && <ItemPriceNormal>{nPrice}원</ItemPriceNormal>}
+		</>
+	) : (
+		<ItemPriceSmall>{sPrice ? sPrice : nPrice}원</ItemPriceSmall>
+	);
+}
+
+export default ItemPrice;
+
 const ItemPriceBlock = styled.span`
 	font-size: ${theme.fontSize.larger};
 	font-weight: Bold;
 	margin-right: 10px;
 `;
+const ItemPriceSmall = styled.div`
+	font-size: ${theme.fontSize.small};
+	font-weight: Bold;
+`;
 const ItemPriceNormal = styled.span`
 	text-decoration: line-through;
 	color: ${theme.colors.grey_text};
 `;
-// const priceComma = (price) => {
-// 	if(price.length<4) return;
-// 	while(price.length>0){
-// 		price.slice(price.length-1)
-// 	}
-// };
-// priceComma('3000');
-function ItemPrice({ sPrice, nPrice }) {
-	return (
-		<>
-			<ItemPriceBlock>{sPrice ? sPrice : nPrice}원</ItemPriceBlock>
-			{sPrice && <ItemPriceNormal>{nPrice}원</ItemPriceNormal>}
-		</>
-	);
-}
-
-export default ItemPrice;
+const priceComma = (price) => {
+	return price.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
+};
