--- frontend/src/components/category/Carousel.md
@@ -1,1 +1,0 @@
-effect : ease , linear, ease-in, ease-out, ease-in-out
