+++ frontend/src/components/mainmenu/MainMenu.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { theme, AlignTextCenter } from '../style/Theme';
-import ItemCard from '../ItemCard';
+import ItemCard from '../card/ItemCard';
 import useFetch from '../useFetch';
 import Title from '../atomic/Title';
 import { useState } from 'react';
@@ -8,14 +8,11 @@ import Loading from '../state/Loading';
 
 function MainMenu() {
 	const basicUrl = process.env.REACT_APP_API_URL + 'best/';
-	//5개: 탭 전체 데이터 요청
 	const [bestDishMenu, bestDishLoading] = useFetch(basicUrl, 'get');
 	const [clickedID, setClickedID] = useState(1);
-	//초기 베스트메뉴 url 설정
 	const [fetchData, setFetchData] = useState(basicUrl + 1);
-	//3개: 초기 베스트메뉴 데이터 요청
 	const [bestData, loadingState] = useFetch(fetchData, 'get');
-	//클릭한 후 해당 탭 데이터 요청
+
 	const handleClick = (target, id) => {
 		setClickedID(id);
 		setFetchData(basicUrl + id);
